requirements.txt
openai==1.55.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
clang
